Refactor default implementation to `verlib2.packaging`
