+++ verlib2/packaging/_structures.py
